various updates to contacts & skillsFluency routes

+++ backend/routes/api/contactsRoutes.js
@@ -6,9 +6,23 @@ const db = require('../../database/db-connector');
 // GET route should basically return all of the contacts that a user has
 router.get('/', async (req, res) => {
     const user_id = req.user.user_id;
+    const contact_id = req.params.contact_id
+    const job_id = req.params.job_id;
     try {
-        const contacts = await db.query('SELECT * FROM contacts WHERE user_id = ?', [user_id]);
-        res.send(contacts[0]);
+        // Fetch all contacts for the authenticated user
+        const contactsQuery = await db.query('SELECT * FROM contacts WHERE user_id = ?', [user_id]);
+        const contactsJobsQuery = await db.query('SELECT * FROM jobs_contacts WHERE contact_id = ?', [contact_id]);
+        
+         // Combine data from contacts and jobs_contacts tables into a single array of contacts objects
+        const contacts = contactsQuery[0].map(contact => {
+            // Find the corresponding contact data based on contact_id
+            const userContact = contactsJobsQuery[0].find(us => us.contact_id === contact.contact_id);
+            return {
+                ...contact,
+                relationship_type: userContact ? userContact.relationship_type : null
+            };
+        });
+        res.status(200).send(contacts);
     } catch (err) {
         console.error('Error getting contacts:', err)
         res.status(500).send('Query error');
@@ -19,9 +33,10 @@ router.get('/', async (req, res) => {
 // POST route should receive data payload from frontend and store in DB
 router.post('/', async (req, res) => {
     const user_id = req.user.user_id;
-    const { first_name, last_name, email, phone, position, notes } = req.body;
+    const { first_name, last_name, email, phone, position, notes, relationship_type } = req.body;
     try {
-        await db.query('INSERT INTO jobs (first_name, last_name, email, phone, position, notes) VALUES (?, ?, ?, ?, ?, ?)', [
+        // Insert new contact into the database
+        await db.query('INSERT INTO contacts (user_id, first_name, last_name, email, phone, position, notes) VALUES (?, ?, ?, ?, ?, ?)', [
             user_id,
             first_name,
             last_name,
@@ -30,6 +45,8 @@ router.post('/', async (req, res) => {
             position || null,
             notes || null,
         ]);
+        // Insert relationship_type
+        await db.query('INSERT INTO jobs_contacts (job_id, contact_id, relationship_type) VALUES (?, ?, ?)', [job_id, contact_id, relationship_type || null]);
         res.status(201).send('Insert succeeded');
     } catch (err) {
         console.error('Error while adding a new contact:', err);
@@ -41,9 +58,12 @@ router.post('/', async (req, res) => {
 // PUT route should receive data payload and update DB for that contact ID
 router.put('/', async (req, res) => {
     const user_id = req.user.user_id;
-    const { contact_id, first_name, last_name, email, phone, position, notes } = req.body;
+    const contact_id = req.params.contact_id
+    const job_id = req.params.job_id;
+    const { first_name, last_name, email, phone, position, notes, relationship_type } = req.body;
     try {
-        await db.query('UPDATE jobs SET first_name = ?, last_name = ?, email = ?, phone = ?, position = ?, notes = ? WHERE contact_id = ? AND user_id = ?',  [
+        // Update contacts
+        await db.query('UPDATE contacts SET first_name = ?, last_name = ?, email = ?, phone = ?, position = ?, notes = ? WHERE contact_id = ? AND user_id = ?',  [
             first_name,
             last_name,
             email || null,
@@ -52,6 +72,12 @@ router.put('/', async (req, res) => {
             notes || null,
             contact_id,
             user_id,
+        ]);
+        // Update jobs_contacts
+        await db.query('UPDATE jobs_contacts SET relationship_type = ? WHERE job_id = ? AND contact_id = ?', [
+            relationship_type || null, 
+            job_id, 
+            contact_id,
         ]);
         res.status(204).send('Update succeeded');
     } catch (err) {
@@ -63,10 +89,14 @@ router.put('/', async (req, res) => {
 // DELETE the specificed resource
 // DELETE route should delete row from DB for that contact ID
 router.delete('/', async (req, res) => {
-    const { contact_id } = req.body;
+    const user_id = req.user.user_id;
+    const contact_id = req.params.contact_id;
+    const job_id = req.params.job_id;
     try {
-        await db.query('DELETE FROM contacts WHERE contact_id = ?', [contact_id]);
-        res.send(204).send('Delete succeeded');
+        // Delete contact and associated jobs_contacts data
+        await db.query('DELETE FROM contacts WHERE contact_id = ? AND user_id = ?', [contact_id, user_id]);
+        await db.query('DELETE FROM job_contacts WHERE contact_id = ? AND job_id = ?', [contact_id, job_id]);
+        res.send(204).send('Contact deleted successfully');
     } catch (err) {
         console.error('Error while deleting contact:', err);
         res.status(500).send('Delete failed');
